test(std/wasi): seperate test runner into a module (#7191)
